\name{xtensor-package}
\alias{xtensor-package}
\alias{xtensor}
\docType{package}
\title{
  R bindings for the xtensor C++ multi-dimensional array library.
}
\description{
  xtensor is a C++ library for multi-dimensional arrays enabling numpy-style broadcasting and lazy computing.
  xtensor-r enables inplace use of R arrays in C++ with all the benefits from xtensor

  \itemize{
    \item C++ universal functions and broadcasting
    \item  STL - compliant APIs.
    \item A broad coverage of numpy APIs (see the numpy to xtensor cheat sheet at \url{http://xtensor.readthedocs.io/en/latest/numpy.html}).
  }

  xtensor-r can be used either to author C++ extensions for R with Rcpp, or applications that embed the R interpreter with RInside.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Sylvain Corlay, Wolf Vollprecht, Johan Mabille

Maintainer: Sylvain Corlay <sylvain.corlay@gmail.com>
}
\keyword{numpy}
\keyword{tensor}
\seealso{
  Development for \pkg{xtensor} can be followed via the GitHub repository
  at \url{http://github.com/QuantStack/xtensor}.
}
